Add balanc0r filter.
Balanc0r is a white/color balance filter taken from the white balance portion
of my Kino Levels filter, which is based on the digikam white balance plugin.
It works by setting the color parameter to what should be a neutral color (but
is not), and the filter determines how to adjust it and all the other colors
to make the image look correct. There is also a parameter to adjust the green
tint since the color temperature part mainly just affects the red and blue
channels. It is most handy when the user interface offers an eye-dropper-like
tool to let the user pick a "white" spot on a displayed frame of the
unfiltered video.

Signed-off-by: Dan Dennedy <dan@dennedy.org>