build: Add CMake build files
This adds CMakeLists.txt files for all plugins. The motivation for this is to enable building of the plugins within Microsoft Visual Studio. At present plugins that depend on OpenCV or gavl won't build.

While the CMake files can be used on *nix to generate automake build files, at the present time they will not be as feature complete as the build system already in place.

MSVC by default doesn't export functions. A module definition file (DEF) has been created for versions 1.0 and 1.1 which enables exporting of functions under MSVC.