CMake build fixes.
Not sure what's the status of the CMake build system, but it has
problems finding the OpenCV library. Below is a patch using pkg-config
instead of the strange stuff used.

This also highlighted the fact that the facebl0r effect has no
CMakeLists.txt in it's directory, a proposal for that file is also just
below.