Fix array out-of-bounds error in partik0l.
When using rand(), it may make sense to use RAND_MAX. But in a build
made on MinGW RAND_MAX is equivalent to a SHRT_MAX, which results in
crashes in blossom_recal() due to accessing the prime array outside of
its bounds. INT_MAX was chosen because on Linux, RAND_MAX is equivalent
to INT_MAX, and UINT_MAX gives different results visually.