CMakeLists: Don't hardcode pkgconfig install dir
Currently the pkgconfig files are hardcoded to be installed to
`$CMAKE_INSTALL_PREFIX/lib/pkgconfig`, which affects installations that
are using a differente library directory. This patch allows the use of
the GNUInstallDirs package to modify the installation libdir at build
time and to install the pkgconfig files to
`$CMAKE_INSTALL_LIBDIR`/pkgconfig.

Addresses #79

A previous attempt was made in #80, which failed without the
GNUInstallDirs package being required.