Speed up cairoblend by reducing memory copies
Instead of premultiplying the destination buffer in-place and using
cairo to copy the destination to the output buffer, premultiply the
destination directly to the output buffer and skip the first cairo
paint.

This reduces the time to blend two 4K video frames processed in parallel
across four cores from 16.8 ms to 13.7 ms (and reduces memory bandwidth
usage if you're constrained by that).