Add pixel-sorting filter
s0rter is a filter that sorts each frame's pixels by hue, saturation,
and lightness (in order of decreasing priority), producing a
rainbow-like effect that's especially interesting for frames of
animated content with a limited color palette.  While it's not
configurable in this iteration, it can easily be extended to support
arbitray hue/saturation/lightness ordering.